+++ lib/binary_resource.dart
@@ -1,0 +1,5 @@
+library binary_marshalling.binary_resources;
+
+import "package:binary_types/binary_types.dart";
+
+part "src/binary_resources/binary_resources.dart";