+++ lib/src/annotations/annotations.dart
@@ -1,5 +1,23 @@
 part of binary_marshalling.annotations;
 
+class CheckBufferSize {
+  final String buffer;
+
+  final String message;
+
+  final String size;
+
+  const CheckBufferSize({this.buffer, this.message, this.size});
+}
+
+class CheckResource {
+  final String message;
+
+  final String name;
+
+  const CheckResource({this.message, this.name});
+}
+
 class NativeName {
   final String name;
 
@@ -10,4 +28,20 @@ class NullTerminated {
   final bool includeNull;
 
   const NullTerminated({this.includeNull: false});
+}
+
+class RegisterResource {
+  final String message;
+
+  final String name;
+
+  const RegisterResource({this.message, this.name});
+}
+
+class UnregisterResource {
+  final String message;
+
+  final String name;
+
+  const UnregisterResource({this.message, this.name});
 }